Refactor pkg/generate/app
