+++ pkg/generate/app/cmd/newapp.go
@@ -3,9 +3,6 @@ package cmd
 import (
 	"fmt"
 	"io"
-	"regexp"
-	"strconv"
-	"strings"
 
 	"github.com/fsouza/go-dockerclient"
 	"github.com/golang/glog"
@@ -17,7 +14,6 @@ import (
 	"k8s.io/kubernetes/pkg/kubectl/resource"
 	"k8s.io/kubernetes/pkg/runtime"
 	"k8s.io/kubernetes/pkg/util/errors"
-	kvalidation "k8s.io/kubernetes/pkg/util/validation"
 
 	authapi "github.com/openshift/origin/pkg/authorization/api"
 	buildapi "github.com/openshift/origin/pkg/build/api"
@@ -27,11 +23,9 @@ import (
 	"github.com/openshift/origin/pkg/generate/app"
 	"github.com/openshift/origin/pkg/generate/dockerfile"
 	"github.com/openshift/origin/pkg/generate/source"
-	imageapi "github.com/openshift/origin/pkg/image/api"
 	"github.com/openshift/origin/pkg/template"
 	outil "github.com/openshift/origin/pkg/util"
 	dockerfileutil "github.com/openshift/origin/pkg/util/docker/dockerfile"
-	"github.com/openshift/origin/pkg/util/namer"
 )
 
 const (
@@ -45,9 +39,6 @@ const (
 // ErrNoDockerfileDetected is the error returned when the requested build strategy is Docker
 // and no Dockerfile is detected in the repository.
 var ErrNoDockerfileDetected = fmt.Errorf("No Dockerfile was found in the repository and the requested build strategy is 'docker'")
-
-// the opposite of kvalidation.DNS1123LabelFmt
-var invalidNameCharactersRegexp = regexp.MustCompile("[^-a-z0-9]")
 
 // AppConfig contains all the necessary configuration for an application
 type AppConfig struct {
@@ -589,138 +580,42 @@ func (c *AppConfig) validateEnforcedName() error {
 	return nil
 }
 
-func ensureValidUniqueName(names map[string]int, name string) (string, error) {
-	// Ensure that name meets length requirements
-	if len(name) < 2 {
-		return "", fmt.Errorf("invalid name: %s", name)
-	}
-
-	// Make all names lowercase
-	name = strings.ToLower(name)
-
-	// Remove everything except [-0-9a-z]
-	name = invalidNameCharactersRegexp.ReplaceAllString(name, "")
-
-	// Remove leading hyphen(s) that may be introduced by the previous step
-	name = strings.TrimLeft(name, "-")
-
-	if len(name) > kvalidation.DNS1123SubdomainMaxLength {
-		glog.V(4).Infof("Trimming %s to maximum allowable length (%d)\n", name, kvalidation.DNS1123SubdomainMaxLength)
-		name = name[:kvalidation.DNS1123SubdomainMaxLength]
-	}
-
-	count, existing := names[name]
-	if !existing {
-		names[name] = 0
-		return name, nil
-	}
-	count++
-	names[name] = count
-	newName := namer.GetName(name, strconv.Itoa(count), kvalidation.DNS1123SubdomainMaxLength)
-	return newName, nil
-}
-
 // buildPipelines converts a set of resolved, valid references into pipelines.
 func (c *AppConfig) buildPipelines(components app.ComponentReferences, environment app.Environment) (app.PipelineGroup, error) {
 	pipelines := app.PipelineGroup{}
-	names := map[string]int{}
+	pipelineBuilder := app.NewPipelineBuilder(c.Name, c.GetBuildEnvironment(environment), c.OutputDocker)
 	for _, group := range components.Group() {
 		glog.V(4).Infof("found group: %#v", group)
 		common := app.PipelineGroup{}
 		for _, ref := range group {
 			refInput := ref.Input()
-
-			var pipeline *app.Pipeline
-			var name string
-
+			from := refInput.String()
+			var (
+				pipeline *app.Pipeline
+				err      error
+			)
 			if refInput.ExpectToBuild {
 				glog.V(4).Infof("will use %q as the base image for a source build of %q", ref, refInput.Uses)
-				input, err := app.InputImageFromMatch(refInput.ResolvedMatch)
-				if err != nil {
+				if pipeline, err = pipelineBuilder.NewBuildPipeline(from, refInput.ResolvedMatch, refInput.Uses); err != nil {
 					return nil, fmt.Errorf("can't build %q: %v", refInput, err)
 				}
-				if !input.AsImageStream {
-					glog.Warningf("Could not find an image stream match for %q. Make sure that a Docker image with that tag is available on the node for the build to succeed.", refInput.ResolvedMatch.Value)
-				}
-
-				strategy, source, err := app.StrategyAndSourceForRepository(refInput.Uses, input)
-				if err != nil {
-					return nil, fmt.Errorf("can't build %q: %v", refInput, err)
-				}
-
-				// Override resource names from the cli
-				name = c.Name
-				if len(name) == 0 {
-					var ok bool
-					name, ok = (app.NameSuggestions{source, input}).SuggestName()
-					if !ok {
-						return nil, fmt.Errorf("can't suggest a valid name, please specify a name with --name")
-					}
-				}
-				name, err = ensureValidUniqueName(names, name)
-				source.Name = name
-				if err != nil {
-					return nil, err
-				}
-
-				// Append any exposed ports from Dockerfile to input image
-				if refInput.Uses.IsDockerBuild() && refInput.Uses.Info() != nil {
-					node := refInput.Uses.Info().Dockerfile.AST()
-					ports := dockerfileutil.LastExposedPorts(node)
-					if len(ports) > 0 {
-						if input.Info == nil {
-							input.Info = &imageapi.DockerImage{
-								Config: &imageapi.DockerConfig{},
-							}
-						}
-						input.Info.Config.ExposedPorts = map[string]struct{}{}
-						for _, p := range ports {
-							input.Info.Config.ExposedPorts[p] = struct{}{}
-						}
-					}
-				}
-				if pipeline, err = app.NewBuildPipeline(refInput.String(), input, c.OutputDocker, strategy, c.GetBuildEnvironment(environment), source); err != nil {
-					return nil, fmt.Errorf("can't build %q: %v", refInput, err)
-				}
-
 			} else {
 				glog.V(4).Infof("will include %q", ref)
-				input, err := app.InputImageFromMatch(refInput.ResolvedMatch)
-				if err != nil {
-					return nil, fmt.Errorf("can't include %q: %v", refInput, err)
-				}
-				name = c.Name
-				if len(name) == 0 {
-					var ok bool
-					name, ok = input.SuggestName()
-					if !ok {
-						return nil, fmt.Errorf("can't suggest a valid name, please specify a name with --name")
-					}
-				}
-				name, err = ensureValidUniqueName(names, name)
-				if err != nil {
-					return nil, err
-				}
-				input.ObjectName = name
-				if pipeline, err = app.NewImagePipeline(refInput.String(), input); err != nil {
+				if pipeline, err = pipelineBuilder.NewImagePipeline(from, refInput.ResolvedMatch); err != nil {
 					return nil, fmt.Errorf("can't include %q: %v", refInput, err)
 				}
 			}
-
 			if c.Deploy {
-				if err := pipeline.NeedsDeployment(environment, c.Labels, name); err != nil {
+				if err := pipeline.NeedsDeployment(environment, c.Labels); err != nil {
 					return nil, fmt.Errorf("can't set up a deployment for %q: %v", refInput, err)
 				}
 			}
 			common = append(common, pipeline)
-
 			if err := common.Reduce(); err != nil {
 				return nil, fmt.Errorf("can't create a pipeline from %s: %v", common, err)
 			}
-
 			describeBuildPipelineWithImage(c.Out, ref, pipeline, c.originNamespace)
 		}
-
 		pipelines = append(pipelines, common...)
 	}
 	return pipelines, nil
@@ -1020,6 +915,9 @@ func (c *AppConfig) run(acceptors app.Acceptors) (*AppResult, error) {
 
 	pipelines, err := c.buildPipelines(imageRefs, env)
 	if err != nil {
+		if err == app.ErrNameRequired {
+			err = fmt.Errorf("can't suggest a valid name, please specify a name with --name")
+		}
 		return nil, err
 	}
