+++ pkg/generate/app/cmd/newapp_test.go
@@ -18,7 +18,6 @@ import (
 	"k8s.io/kubernetes/pkg/kubelet/dockertools"
 	"k8s.io/kubernetes/pkg/runtime"
 	"k8s.io/kubernetes/pkg/util/sets"
-	kvalidation "k8s.io/kubernetes/pkg/util/validation"
 
 	buildapi "github.com/openshift/origin/pkg/build/api"
 	client "github.com/openshift/origin/pkg/client/testclient"
@@ -29,7 +28,6 @@ import (
 	"github.com/openshift/origin/pkg/generate/source"
 	imageapi "github.com/openshift/origin/pkg/image/api"
 	templateapi "github.com/openshift/origin/pkg/template/api"
-	"github.com/openshift/origin/pkg/util/namer"
 )
 
 func skipExternalGit(t *testing.T) {
@@ -1262,72 +1260,6 @@ func TestNewAppBuildConfigEnvVars(t *testing.T) {
 		if !reflect.DeepEqual(test.expected, got) {
 			t.Errorf("%s: unexpected output. Expected: %#v, Got: %#v", test.name, test.expected, got)
 			continue
-		}
-	}
-}
-
-func TestEnsureValidUniqueName(t *testing.T) {
-	chars := []byte("abcdefghijk")
-	longBytes := []byte{}
-	for i := 0; i < (kvalidation.DNS1123SubdomainMaxLength + 20); i++ {
-		longBytes = append(longBytes, chars[i%len(chars)])
-	}
-	longName := string(longBytes)
-	tests := []struct {
-		name        string
-		input       []string
-		expected    []string
-		expectError bool
-	}{
-		{
-			name:     "duplicate names",
-			input:    []string{"one", "two", "three", "one", "one", "two"},
-			expected: []string{"one", "two", "three", "one-1", "one-2", "two-1"},
-		},
-		{
-			name:     "mixed case names",
-			input:    []string{"One", "ONE", "tWo"},
-			expected: []string{"one", "one-1", "two"},
-		},
-		{
-			name:     "non-standard characters",
-			input:    []string{"Emby.One", "test-_test", "_-_", "@-MyRepo"},
-			expected: []string{"embyone", "test-test", "", "myrepo"},
-		},
-		{
-			name:        "short name",
-			input:       []string{"t"},
-			expectError: true,
-		},
-		{
-			name:  "long name",
-			input: []string{longName, longName, longName},
-			expected: []string{longName[:kvalidation.DNS1123SubdomainMaxLength],
-				namer.GetName(longName[:kvalidation.DNS1123SubdomainMaxLength], "1", kvalidation.DNS1123SubdomainMaxLength),
-				namer.GetName(longName[:kvalidation.DNS1123SubdomainMaxLength], "2", kvalidation.DNS1123SubdomainMaxLength),
-			},
-		},
-	}
-
-tests:
-	for _, test := range tests {
-		result := []string{}
-		names := make(map[string]int)
-		for _, i := range test.input {
-			name, err := ensureValidUniqueName(names, i)
-			if err != nil && !test.expectError {
-				t.Errorf("%s: unexpected error: %v", test.name, err)
-			}
-			if err == nil && test.expectError {
-				t.Errorf("%s: did not get an error.", test.name)
-			}
-			if err != nil {
-				continue tests
-			}
-			result = append(result, name)
-		}
-		if !reflect.DeepEqual(result, test.expected) {
-			t.Errorf("%s: unexpected output. Expected: %#v, Got: %#v", test.name, test.expected, result)
 		}
 	}
 }
